:recycle: separate ReportingConfig from InstrumentationConfig
